# Optimize StumpyPNG.write