--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 /.crystal
 /.shards
 /libs
+/lib
 /tmp