--- a/src/stumpy_png.cr
+++ b/src/stumpy_png.cr
@@ -18,38 +18,56 @@ module StumpyPNG
   def self.write(canvas, path)
     datastream = Datastream.new
 
-    ihdr_io = IO::Memory.new(13)
-    ihdr_io.write_bytes(canvas.width, IO::ByteFormat::BigEndian)
-    ihdr_io.write_bytes(canvas.height, IO::ByteFormat::BigEndian)
-    ihdr_io.write_byte 16_u8 # bit depth = 16 bit
-    ihdr_io.write_byte 6_u8  # color_type = rgba
-    ihdr_io.write_byte 0_u8  # compression = deflate
-    ihdr_io.write_byte 0_u8  # filter = adaptive
-    ihdr_io.write_byte 0_u8  # interlacing = none
+    ihdr = uninitialized UInt8[13]
+    ihdr_slice = ihdr.to_slice
+    # TODO: use IO::ByteFormat::BigEndian.encode(canvas.widht, ihdr_slice) once 0.20.1 is out
+    encode(canvas.width, ihdr_slice)
+    encode(canvas.height, ihdr_slice + 4)
+    ihdr_slice[8] = 16_u8 # bit depth = 16 bit
+    ihdr_slice[9] = 6_u8  # color_type = rgba
+    ihdr_slice[10] = 0_u8 # compression = deflate
+    ihdr_slice[11] = 0_u8 # filter = adaptive
+    ihdr_slice[12] = 0_u8 # interlacing = none
 
-    datastream.chunks << Chunk.new("IHDR", ihdr_io.to_slice)
+    # datastream.chunks << Chunk.new("IHDR", ihdr_io.to_slice)
+    datastream.chunks << Chunk.new("IHDR", ihdr_slice)
 
-    buffer = IO::Memory.new
-
+    buffer = Bytes.new(canvas.height * (1 + canvas.width * 8))
+    buffer_ptr = buffer
+    i = 0
     canvas.each_column do |col|
-      buffer.write_byte(0_u8) # filter = none
+      buffer_ptr += 1
       col.each do |pixel|
         {pixel.r, pixel.g, pixel.b, pixel.a}.each do |value|
-          buffer.write_bytes(value, IO::ByteFormat::BigEndian)
+          # TODO: use IO::ByteFormat::BigEndian.encode(value, buffer_ptr) once 0.20.1 is out
+          encode(value, buffer_ptr)
+          buffer_ptr += 2
         end
       end
     end
 
-    # Reset buffer position
-    buffer.pos = 0
-
     compressed = IO::Memory.new
     Zlib::Deflate.new(compressed) do |deflate|
-      IO.copy(buffer, deflate)
+      deflate.write(buffer)
     end
 
     datastream.chunks << Chunk.new("IDAT", compressed.to_slice)
+
     datastream.chunks << Chunk.new("IEND", Bytes.new(0))
     datastream.write(path)
+  end
+
+  private def self.encode(value : UInt16, slice)
+    bytes = pointerof(value).as(UInt8[2]*).value
+    slice[0] = bytes[1]
+    slice[1] = bytes[0]
+  end
+
+  private def self.encode(value : Int32, slice)
+    bytes = pointerof(value).as(UInt8[4]*).value
+    slice[0] = bytes[3]
+    slice[1] = bytes[2]
+    slice[2] = bytes[1]
+    slice[3] = bytes[0]
   end
 end