The main optimization is to avoid IO::Memory and use a fixed slice and write into that, which is a bit faster.

Before:

```
Read: 151.55ms (37.56%)
Transform: 17.8ms (4.41%)
Write: 234.15ms (58.03%)
Total: 403.5ms
```

After:

```
Read: 148.56ms (42.63%)
Transform: 20.02ms (5.74%)
Write: 179.88ms (51.62%)
Total: 348.46ms
```

As a side note, I noted memory consumption is pretty high, mostly because a lot of buffers are needed that hold all of the data in memory, sometimes multiple times (for example data is read, then analyzed in a different buffer). Is there maybe a different way to do it without preloading everything? For example the Python version only uses 17MB while Crystal uses 53MB or so. In any case, that's an optimization that can be done later, and speed is usually more important.

Another question is whether the whole data needs to be written and then deflated/inflated instead of being done lazily... I see some crc32 checks so maybe there's no way around it, or maybe one could use an IO::MultiWriter to write the output and at the same time compute a crc, but I don't know (I think right now there's no way to do that in the standard library)