@l3kn 2016-11-30
Thanks again!

I'll try to think about some ways do fix the memory consumption problem
and take a look at png libraries for other languages,
maybe there is some trick I don't know yet